tests(repair): displace sudo tests into seqtests
